A quick hack to fix an issue that arrises in surface generation of a texture.
Specifically, the generation functions maintain their capture of the
original label it was created for. After a move, new changes in the data
have no affect, which is the best-case scenario. Ideally there would be a
fix in the PpM core that would still allow for efficient move-semantics
for the cache.