New vector structure for our own uses.
Everything in the game namespace now uses it (where possible). The
template operators for the vector are cleaner/just work and aren't
ruined with the ridiculous pong::math:: prefix everywhere. Because of
this change, the moving delta and player position are now
double-floating point values. That means correct player movement has
been implemented. Changing players is currently disabled (commented out)
but that comment is there, ready to be turned into code!

Also more commenting everywhere + more use of auto.