Each player can now have it's own view radius.
Now it can be modified with items for instance! The fog is also
resetable for each player. This alleviates the problem of default
construction for players. It just means without any explicit command,
the player simply has now fog. This gets rid of the weirdness in
Map.cpp in initialization of the six player objects.