New Entity_Data structure.
Contains data used in combat for both the Player and Enemy_Impl structure to
use. The Entity_Impl structure also now stores a defense value used to
propagate the Entity_Data member of the Entity_Instance. The "defense"
property is parsed from the json.

Each "power" property in the json itself has also been changed to
"defense". In addition to that, they and the life properties now
somewhat corrospond to what they may end up being. Basically a rough
sketch of combat balance as of now.