Added the beginnings of a new UI system.
The hiearchy is roughly based off of my broken understanding of Android
UI. That being said, it seems easier to extend compared the last system,
which is most important for the moment. Currently the following is
pretty much implemented.

- Linear layout (works similarly to Android, every child element is
centered in cells determined from the orientation and volume of the
linear layout.
- Label. Rendering works and they are properly loaded from json.
- Sprite. Sprites probably render correctly but aren't loaded from json
yet.

TODO:
 - The view container should have better default implementations of the
 virtual functions.