Layout dispatch now doesn't use a parameter but an inherited member function.
This allows for consistency for storage of both the parent, containing
volume and the volume that is actually rendered, the exact behavior and
usage of this second volume, stored in the View base class is not
defined specifically. But is persisted between layouts (set to the new
return value of the layout_ virtual member function).