New View property 'size' that acts like a user-defined 'minimum_extent'.
Basically, the get_minimum_extent function returns the maximum of the two. If
the user-defined size is too small. The function still returns the view's own
minimum extent requirement. However, if that view doesn't have one (like a Bar)
this will prevent the whole view from being collapsed in a layout, which seems
to be happening.