Combat works now!
The only issue currently is that animations will go and nothing waits for them
to finish. This means they are cut off automatically if the player or the enemy
die and also the enemy will attack the player while it's own health is still
animating. Kind of disorienting; will fix.