Everything now uses the assets items class instead of the old one in decl/
One biggie change is now the Map, for one stores the assets it needs directly
rather than the string references, this will allow us to use them a lot in
different contexts without re-searching through the assets vector each time.
This sacrifixes hot-plugging we might want later, but I figure for things like
images it would be better to just reload the Map. I mean, it's not we could
change the json asset the map was constructed with all that easily anyway. I
figure it's a non-issue right now. The only thing we really need to worry about
is language. After the label can translate it's arg strings on the fly, we
should be able to do that.